fix(architecture): Moving the tsp classes to the service package, and updating the module infos ensuring proper hexagonal architecture compliance.
