+++ application/src/main/java/fr/delivrooom/application/service/TourCalculatorService.java
@@ -1,11 +1,13 @@
 package fr.delivrooom.application.service;
 
 import fr.delivrooom.application.model.*;
-import fr.delivrooom.application.model.tsp.DynamicProgrammingTSP;
-import fr.delivrooom.application.model.tsp.TSP;
-import fr.delivrooom.application.model.tsp.TSP3;
 import fr.delivrooom.application.port.in.CalculateTourUseCase;
 import fr.delivrooom.application.port.out.NotifyTSPProgressToGui;
+import fr.delivrooom.application.service.tourcalculator.CityGraph;
+import fr.delivrooom.application.service.tourcalculator.ShortestPathsGraph;
+import fr.delivrooom.application.service.tourcalculator.tsp.DynamicProgrammingTSP;
+import fr.delivrooom.application.service.tourcalculator.tsp.TSP;
+import fr.delivrooom.application.service.tourcalculator.tsp.TSP3;
 
 import java.util.*;
