+++ application/src/main/java/fr/delivrooom/application/service/tourcalculator/tsp/DynamicProgrammingTSP.java
@@ -1,8 +1,8 @@
-package fr.delivrooom.application.model.tsp;
+package fr.delivrooom.application.service.tourcalculator.tsp;
 
 import fr.delivrooom.application.model.DeliveriesDemand;
-import fr.delivrooom.application.model.Graphe;
 import fr.delivrooom.application.port.out.NotifyTSPProgressToGui;
+import fr.delivrooom.application.service.tourcalculator.Graphe;
 
 import java.util.*;
