+++ application/src/main/java/fr/delivrooom/application/service/tourcalculator/tsp/IteratorSeq.java
@@ -1,6 +1,6 @@
-package fr.delivrooom.application.model.tsp;
+package fr.delivrooom.application.service.tourcalculator.tsp;
 
-import fr.delivrooom.application.model.Graphe;
+import fr.delivrooom.application.service.tourcalculator.Graphe;
 
 import java.util.Collection;
 import java.util.Iterator;
