+++ application/src/main/java/module-info.java
@@ -1,8 +1,11 @@
 module fr.delivrooom.application {
     requires javafx.graphics;
+
+    // Adapters should only see the model classes along with the port interfaces
     exports fr.delivrooom.application.model;
     exports fr.delivrooom.application.port.in;
     exports fr.delivrooom.application.port.out;
-    exports fr.delivrooom.application.service;
 
+    // Services should be available to use only by the bootstrap
+    exports fr.delivrooom.application.service to fr.delivrooom.config;
 }
